Refactoring auto-scaler logic
