+++ core/src/emr-eks-platform/index.ts
@@ -2,4 +2,3 @@ export { EmrEksClusterProps, EmrEksCluster, Autoscaler } from './emr-eks-cluster
 export { EmrEksNodegroupOptions, EmrEksNodegroup } from './emr-eks-nodegroup';
 export { EmrVirtualClusterOptions } from './emr-virtual-cluster';
 export { EmrManagedEndpointOptions } from './emr-managed-endpoint';
-export { KarpenterProvisioner } from './karpenter-provisioner';
