+++ core/test/unit/cdk-nag/nag-emr-eks.test.ts
@@ -14,6 +14,7 @@ import { Annotations, Match } from 'aws-cdk-lib/assertions';
 // eslint-disable-next-line import/no-extraneous-dependencies,import/no-unresolved
 import { AwsSolutionsChecks, NagSuppressions } from 'cdk-nag';
 import { Autoscaler, EmrEksCluster } from '../../../src/emr-eks-platform';
+import { Autoscaler, EmrEksCluster } from '../../../src/emr-eks-platform';
 //import { NotebookPlatform, StudioAuthMode } from '../notebook-platform/';
 
 
@@ -24,7 +25,7 @@ Aspects.of(mockApp).add(new AwsSolutionsChecks());
 
 EmrEksCluster.getOrCreate(stack, {
   eksAdminRoleArn: 'arn:aws:iam::123445678912:role/myRole',
-  autoscaling: Autoscaler.CLUSTER_AUTOSCALER
+  autoScaling: Autoscaler.CLUSTER_AUTOSCALER
 });
 
 NagSuppressions.addResourceSuppressionsByPath(
