+++ core/test/unit/cdk-nag/nag-notebook-platform.test.ts
@@ -14,7 +14,7 @@ import { ManagedPolicy, PolicyStatement } from 'aws-cdk-lib/aws-iam';
 // eslint-disable-next-line import/no-extraneous-dependencies,import/no-unresolved
 import { AwsSolutionsChecks, NagSuppressions } from 'cdk-nag';
 import { SSOIdentityType } from '../../../lib';
-import { Autoscaler, EmrEksCluster, EmrVersion } from '../../../src/emr-eks-platform';
+import { Autoscaler, EmrEksCluster } from '../../../src/emr-eks-platform';
 import { NotebookPlatform, StudioAuthMode } from '../../../src/notebook-platform';
 
 const mockApp = new App();
@@ -24,7 +24,7 @@ Aspects.of(mockApp).add(new AwsSolutionsChecks({ verbose: true }));
 
 const emrEks = EmrEksCluster.getOrCreate(stack, {
   eksAdminRoleArn: 'arn:aws:iam::123445678912:role/gromav',
-  autoscaling: Autoscaler.CLUSTER_AUTOSCALER
+  autoScaling: Autoscaler.CLUSTER_AUTOSCALER
 });
 
 const notebookPlatform = new NotebookPlatform(stack, 'eks-emr-studio', {
